requirements.txt
importlib-metadata==3.10.0
numpy==1.19.5
pydantic==1.8.2
tensorflow==2.4.1
Flask==2.0.1
imageio==2.9.0
opencv-python==4.5.1.48
torch>=1.4.0
torchvision>=0.5.0
scikit-image>=0.16.2
click>=7.0
scipy>=1.4.1
pillow==8.2.0